+++ src/main/java/com/jala/model/criteria/CriteriaConverterVideo.java
@@ -10,7 +10,9 @@
  * with Jala Foundation.
  */
 
-package com.jala.convertor.models;
+package com.jala.model.criteria;
+
+import com.jala.model.criteria.CriteriaConverter;
 
 /**
  * CriteriaConverterVideo
