+++ src/main/java/com/jala/search/controller/ControllerSearch.java
@@ -11,9 +11,9 @@
  */
 package com.jala.search.controller;
 
-import com.jala.search.models.Asset;
-import com.jala.search.models.CriteriaSearch;
-import com.jala.search.models.SearchFile;
+import com.jala.model.assetFile.Asset;
+import com.jala.model.criteria.CriteriaSearch;
+import com.jala.model.SearchFile;
 import com.jala.utils.Logs;
 import com.jala.view.JPanelSearchGral;
 import org.apache.log4j.Logger;
